alex-hf/requirements.txt
# Function dependencies, for example:
# package>=version
requests
google-cloud-secret-manager
google-crc32c
openai
google-cloud-datastore
